Blue green is observed during camera open
When app moves to groundback all the data structure remains in stack
and when app open again , app tries to clear view from cashed data
this is when static camera frames is seen but later app tries to open
with updated surfaces this is when blue screen is observed and blue
screen visible till first camera frames appears.

When app is moving to background we are cleaning all the data by
killing the process so next time camera opens with fresh data.

Tests Done : tested for multiple times camera open\close and no
blue screen is observed.

Tracked-On: OAM-127022
Signed-off-by: shivasku82 <shiva.kumara.rudrappa@intel.com>