multi camera app closes after video playback
After video play from multi camera app its expected to come
back to playing list but multi camera app is getting closed.

As Multi camera application enter to play onStop for multi
camera activity is called which is triggering app exit.This
change is needed only for IASW image where 4 camera is getting
preview hence for BM we are not exiting the app.

Tracked-On: OAM-127692
Signed-off-by: shivasku82 <shiva.kumara.rudrappa@intel.com>